Actually use layouts in the quickdialer applet
While layouts aren't very QML-like, they are integral to the architecture of
plasma, so in plasma applets like this they may need QML bindings. There are
cursory layout bindings in the plasma bindings for QML already, although they
could use to be redesigned for easier use. But positioners are not going to
work for managing sizes like layouts do (although an intermediate level of
layouting is plausible).

Summary: To get layouts-like behaviour, just use layouts. Even in QML.


svn path=/trunk/playground/base/plasma/shells/mobile/; revision=1147803