zero cache misses in my test set
- we need to be more strict about the requested source, otherwise the
  nameing of the cached files changes due to re-conversion to QUrl,
which might add slashes, etc. So store the source separately, and thus
make sure one source always corresponds to one cached file.

This implements the second part of the caching, where we're currently
going through a .png file read from disk, so we can reference these
pictures in the QML code.

Needs cleanup, which follows in the next commit.