Inject intermediate frame
It seems that the performance bottleneck is actually that the signals do
not arrive fast enough to paint smoothly. By triggering the update from
the QML applet a bit more often, we get less lag in the painting, a more
reasonable framerate (> 20 on my machine, from about half of that).

This can probably be made more efficient, but it's interesting to see
that this patch actually brings an improvement in perceived performance
by doing *more* work. Quite surprising.