we can add new bookmarks now
First try to add bookmarks from the webdashboard applet/containment.

There's a new Item, containing an icon and a lineedit which connects to
the org.kde.active.bookmarks engine's "add:<URL>" source, which then
adds the bookmark.

It's a bit wonky, as it doesn't properly update the source right now,
but basically is functional. Which is quite cool. :)