Use KUriFilter to mangle input
This makes all the nice webshortcuts work, which is especially cool
given that input on a keyboard sucks. Also, just entering something
searches now. Good stuff.

The mechanism is as follows:
- user enters a fuzzy "url" or webshortcut
- the UrlInput Item signals urlEntered
- View::onUrlEntered filters the input and
- sets the result as property "filteredUrl" on the UrlInput Item
- the UrlInput Item reacts to this change by loading the filtered URL
  using a new function in webbrowser.qml

The filter is also applied to a command-line  argument passed into
rekonq-active, so can be used from "outside" as well.