streamline activity switching, and fix initial position of recommendations panel
Q: why does it take a few seconds to switch activities?
A: because there was a 2s pause animation and then the activity was scaled in(!)

Q: what happens when that is removed?
A: things get faster.

:)

i left the actual states in place as we may wish to later re-instate some sort of visual effect.
however, as they are completely non-visible (probably due to framerate possibilities on the tablets;
in fact, i only noticed this scaling in on my laptop!), i don't see a point to them right now.