partial workaround to the popups eating events
when a popup window is visible: a click anywhere hides it: but the old qgraphicswidget will continue to think it's under the mouse
doesn't seem to be any good way to properly reset this.
this msolution will still caused a missed click after the popup is gone, but gets the situation unblocked.

anybody has a better idea on this?
CCMAIL:active@kde.org