Implement blacklist in config to not show apps
This is in preparations for Plasma Active One, where we do not want
to show everything, its dog, and the fleas on the dog's tail.

A comprehensive list of blacklisted apps is going into /config/.

Same is there for apps as well.

The config file looks like this:

[blacklist]
apps=minesweeper,dolphing,gimp,
plasmoids=org.kde.active.nonsense,org.kde.active.nonnonsense,org.kde.inactive.noplasmoid,

and so forth.