Cleaned up widgets blacklist, tweaks to apps
Tweaks to apps have been done according to the discussion on the
mailinglist.

I've done a sweep through the applets I've installed (workspace,
plasma-mobile and contour, plasma-addons) and added only those that are
not useful, or usable in this state. It leaves us with a nice collection
of interesting plasmoids.

If you're missing a plasmoid in there that works well, or see one that
does not, please reply to the list, preferably with the plugin name you'd
like to see added or removed.

CCMAIL:active@kde.org