make it somehow work
We have a binary /usr/bin/active-firstrun which waits for a bit, then
reads the configured activities, which are created on first start of
plasma-device. For preconfigured activities (chosen by name!), we add a
few bookmarks and apps on activity screen. This needs to cleaned this
list up with fania's specs, btw).

Caveats:
- waits for 60 seconds right now in order to have nepomuk start up,
  otherwise soprano can't find the server
- after first run, bookmarks and apps appear on all activities. this
  seems to 'fix itself' after restarting plasma-device. Maybe the
  activityscreen doesn't pay close enough attention when bookmarks / apps
  are added?
- For bookmarks, we don't get previews: bookmarks are likely to be added
  by active-firstrun before the network is up. Should probably show a
  suitable icon in that case and update the preview later on.
- .desktop file in /config/ coming up

I'm not overly happy with this, but it might be good enough for PA1, as
it makes the UI quite a bit more discoverable. What do you guys think?

CC:active@kde.org