Disable webthumbnails: crashes
I'm getting regular crashes after the preview has been created by
qtwebkit, disable the thumbnails for now to not bring plasma down.

This should probably move to a plugin for KDE's preview / thumbnailer,
as that would happen out of process, so doesn't involve blocking plasma
while the preview is being rendered in the background by webkit.

For now: slightly less fancy, but not crashing and not blocking