Remove image cache and webthumbnailer
We use the caching mechanism of KIO::PreviewJob, and the new
webthumbnailer that's going into KDE runtime.

The webthumbnailing had been disabled anyway already, so that's no
functional change. The caching moves to PreviewJob, where it's done
anyway. The resulting binary on my disk is 587K vs 982K, nice save for
something that's now done dynamically.