Add readEntry method to ConfigModel
This makes it a bit redundant, but the model handling is not really
suited for key value storage. So it'll be going one way, or the other.

The ConfigTest module has been reworked to write and then read back
valaues to the config, and fails epically in some cases. Putting a date
into KConfig add months, for example. String comparison usually works,
bool fails, and so on. Needs more work, but at least now problems are
visible.