Better support for KWin's TabBox feature
TabBox is now completely controlled from Plasma side. It gets
activated when the dragging starts and connects to the item selected
signal. When an item got selected the panel is closed. As soon as
the panel closed either accept or reject of TabBox is invoked to
close it and (if an item was selected) to activate the window.

This means the switching is slightly delayed.