Make the pressed animation not "hang"
I'm still not 100% happy with this interaction. We want immediate
feedback when the item is pressed, preferably not when it's flicked, but
we have no way to make sure we won't flick on press, this would take a
bit of time.

Anyway, behaviour is a bit less weird with this change.
	../../plasma-mobile.kdev4