Fix delete activity dialog.
I tried to make the dialog's opacity always 1 but with three or less
activities it always gets below 1 if the activity is not close to the
PathView center. It is still readable though. With four activities and
beyond the opacity stays at 1.

The dialog will be oversized if you slide the activity thumbnail to the
PathView center. Well, we can say it is to improve accessibility :-)

CCBUG: 288426