Introduce SettingsItem
SettingsItem encapsulates the loading and switching of a module. It
holds a pagestack used to switch module Items, the Plasma.Package to
load the module's QML package, and the SettingsComponent which takes
care of loading the C++ extensions belonging to a module.

SettingsItem is also the "main API" for loading Active Settings dialogs.