Show Home button conditionally
based on what the devicecapabilities engine reports. This means that by
adding a configfile device-capabilitiesrc with:

[Input]
hasHomeButton=true

You'll hide the home button. It still does take space in the panel, not
sure what the desired solution here would be. Move SLC to the right
then?

CCMAIL:aseigo@kde.org