Blacklist for applicationsmodel
This allows hiding applications on the homescreen. It's useful because
everybody and her dog puts .desktop files into /usr/share/applications,
and many of them are not at all useful. We'll try to do some more
elaborate filtering (watch the next commits), but this provides a
baseline to really hide some things that we don't want to show.

Reviewed-by: notmart