Ensure Wayland integration in FullScreenPanel is setup before window gets created
Perform the Wayland integration directly when the window gets created and
not just on the show event. Show event is too late as the compositor
performs already checks on it like sizing changes.

During the setup we call the QWindow::create method to ensure that the
low-level wayland surface is created.