Don't show empty pages on the right
We want to show as many of our pages as possible, so if we scroll to a page on
the stack make sure that we don't show empty pages on the right just because we
show that page on the left edge.

This issue only has come up with Qt 5.6.0 RC so I wonder if I'm fixing a
symptom instead of the actual problem as with earlier Qt versions scrolling did
the right thing.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>