Remove clone of battery monitor from plasma-phone-components
Summary:
Clone of battery monitor was done for two reasons,

- Showing only single battery instead of 3 dummy
- Simplified user interface instead of one in desktop

However upon investigation battery monitor showing 3 dummy battery is
possibly bug in upower, and those needs to be blacklisted in either,
powermanagement dataengine or battery monitor.

And as for simplifying the user interface, I guess desktop battery
monitor should adapt based on platform it is being used in.

Reviewers: broulik, mart, #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: plasma-devel

Differential Revision: https://phabricator.kde.org/D2411