Disable the startup sound
This served us well so far, the start-up sound was added as debug
mechanism when GUI was not running on Nexus 5 in early days. However now
that we have better mechansims to debug GUI (journalctl, UART, ssh etc).
This no-longer is needed. In addition to that we got enough complaints
about the startup sound if people are trying to hack on plasma mobile on
quiet place like library.

As a historical thing, sound file is still present, just not running
anymore.