homescreen: Use the generic name for KDE Applications
Especially fir newcomers, it can be a bit complex to see that Index is a
File Manager, that Genview or Koko are to be used to browse pictures or
Alligator for following RSS feeds.

This change makes all applications that start as org.kde. to be shown in
their generic name (File Manager, Image Viewer, Feed Reader), while
leaving 3rd party projects