Remove the krunner widget from home screen by default
Because the widget does not have the text entry internally, and pops
up a new window for entering search, which looks confusing and odd on
the phone, we just don't add it by default for now until this can be
fixed. This makes the first impression a bit better, and hides the
annoyance behind needing to manually add the widget. Once the widget
is fixed to not behave in such an odd manner, we can evaluate adding
it back to the default layout.