power menu: New power menu design
This MR introduces a redesigned power menu ("logout" look-and-feel). The design has been changed to (imo) be more inline with the rest of the recent Plasma Mobile design changes, and a bit more professional (ex. removing the scale animation in favor of just a fade). 

Design suggestions are greatly appreciated!

<img src="/uploads/8534b89984ca24abb163ac04cfff6d8d/image.png" alt="New Power Menu Design" width="40%"/>


Comparison (left is new design, right is the current one):

<img src="/uploads/5fa75374e2b01ca5bb90e2c44ef86510/image.png" width="40%"/>