homescreens/folio: Make it easier to click on delegate context menus without starting drag
Dragging delegates on the homescreen involves holding onto it until the context menu pops up, and then dragging.

We currently suppress swipe detection (physically swipe a few pixels before the swipe gets registered) when dragging a delegate so that dragging can start immediately. However, this doesn't make sense if a context menu gets involved, because it can be hard to "let go" to tap the context menu without accidentally triggering the drag action and closing the context menu.

This MR re-enables swipe detection if a context menu is involved with delegate dragging.