shell: Sync implementations with Plasma Desktop
This fixes some of the oddities with applets that we experience that are designed for Desktop. We don't really have a need for different behaviour from Desktop, so it's probably best to follow it for the least amount of bugs.

Eventually we perhaps should derive our shell package from desktop?