quicksettings/docked: Switch navigation panel with desktop panel in docked mode
This implements panel switching such that in docked mode, the navigation panel gets replaced by a desktop panel. If Plasma Desktop is not installed, it falls back to the mobile navigation panel.

This is also useful for tablets to be used more like a laptop when a keyboard is attached.

Current caveats:
- There currently doesn't not seem to be API in Plasma to set the screen a panel is on, so I can't place the panel on an external display
- The desktop panel settings aren't saved when docked mode is disabled, and so the task manager settings are wiped