homescreens/halcyon: Fix favorites delegate not activating with touch
BUG: 486554

In Qt6 a bunch of weird behavior started happening with handlers which broke a bunch of things only in touch.

Remove the HoverHandler here which apparently wasn't used for anything, and causes any MouseArea to stop sending clicked and pressAndHold signals. DragHandler also appears to have stopped working with touch (for the folder close swipe gesture), but that involves far more rewriting to replace...