[shell]: locktouchscreen from quicksettings
This changeset contains a mechanism to temporarily disable the
touchscreen while making it easy to re-enable and see-through.

The purpose is to provide a mechanism that allows avoiding mistouches
when the device is in use.

To achieve this:
- we're adding a button in the Quicksettings drawer "Lock Touchscreen"
  that will open a full-screen, transparent window
- it starts with an animation telling "Touchscreen Locked", then fades
  out
- when touched, an overlay shows a slider saying "Slide to Unlock",
  which fades out again after a short timeout (this was a "mistouch")
- when the user wants to enable the touchscreen again, she can slide the
  slider all the way to the left, which will show a short animation
  "Touchscreen Unlocked" and then hide (destroy) this window

The window is created and destroyed dynamically to reduce memory
consumption while it's not being used.

Signed-off-by: Sebastian Kügler <sebas@kde.org>