hapticsplugin: Port to feedbackd
This is an initial port to feedbackd for the haptics plugin.

This implementation is a simple port to have the motor enabled for a certain duration. We will eventually want to use feedbackd events to trigger these instead.

Related MR for qtfeedback: https://invent.kde.org/jbbgameich/ktactilefeedback/-/merge_requests/2

https://invent.kde.org/teams/plasma-mobile/issues/-/issues/10