new window management scripts
This changeset reworks Plasma Mobile's window management.

Instead of relying on kwinrc setting (and changing) policies for all
windows, we now use a (new) javascript callback into kwin to allow for
more fine-grained window management. This for example, allows us to
avoid showing modal dialogs fullscreen for certain apps.

This changeset implements more fine-grained window management using a
javascript kwin script and the placement callback in this MR:
https://invent.kde.org/plasma/kwin/-/merge_requests/8079

Eventually, this MR supersedes the convergentwindows script and replaces
it with two separate JavaScripts:
- mobilewindows: used when non-docked
- dockedwindows: used when docked

The scripts loop over open windows when loaded and change to border
shown and maximize when mobile, at least for non-modal windows.

There's a change to plasma-mobile-envmanager here which makes the
plasmashell load and unload the new mobilewindows and dockedwindows
winscript when switching between "mobile" and convergenceMode.
It also removes the Window section (may need some code for migration!).

Signed-off-by: Sebastian Kügler <sebas@kde.org>